feat(sign): Create tags.js
Create tags.js to delete duplicated const in signin, signup.js
